Move globaltracking::RecoContainer into a separate data formats library
