+++ DataFormats/Detectors/CMakeLists.txt
@@ -23,6 +23,7 @@ add_subdirectory(TRD)
 add_subdirectory(PHOS)
 add_subdirectory(CPV)
 add_subdirectory(CTP)
+add_subdirectory(GlobalTracking)
 if(ENABLE_UPGRADES)
   add_subdirectory(Upgrades)
 else()
