+++ Detectors/Vertexing/include/DetectorsVertexing/SVertexer.h
@@ -15,7 +15,7 @@
 #define O2_S_VERTEXER_H
 
 #include "gsl/span"
-#include "GlobalTracking/RecoContainer.h"
+#include "DataFormatsGlobalTracking/RecoContainer.h"
 #include "ReconstructionDataFormats/PrimaryVertex.h"
 #include "ReconstructionDataFormats/V0.h"
 #include "ReconstructionDataFormats/Cascade.h"
